Clean up typings
