+++ packages/web-speech-cognitive-services/src/SpeechServices/SpeechToText/SpeechRecognitionEvent.ts
@@ -54,4 +54,8 @@ export default class SpeechRecognitionEvent<
   get results(): SpeechRecognitionResultList {
     return this.#results;
   }
+
+  override get type(): T {
+    return super.type as T;
+  }
 }
