+++ packages/web-speech-cognitive-services/src/SpeechServices/SpeechToText/SpeechRecognitionResult.ts
@@ -1,4 +1,4 @@
-import FakeArray from './FakeArray';
+import FakeArray from './private/FakeArray';
 
 export type SpeechRecognitionResultInit = {
   isFinal: boolean;
