+++ packages/web-speech-cognitive-services/src/SpeechServices/SpeechToText/cognitiveServicesAsyncToPromise.ts
@@ -5,7 +5,7 @@ export default function cognitiveServicesAsyncToPromise<
     resolve: (returnValue: R) => void,
     reject: (error: unknown) => void
   ) => void
->(fn: T): () => Promise<R>;
+>(fn: T, context?: undefined | unknown): () => Promise<R>;
 
 export default function cognitiveServicesAsyncToPromise<
   R,
@@ -15,7 +15,7 @@ export default function cognitiveServicesAsyncToPromise<
     resolve: (returnValue: R) => void,
     reject: (error: unknown) => void
   ) => void
->(fn: T): (arg0: P0) => Promise<R>;
+>(fn: T, context?: undefined | unknown): (arg0: P0) => Promise<R>;
 
 export default function cognitiveServicesAsyncToPromise<
   R,
@@ -27,7 +27,7 @@ export default function cognitiveServicesAsyncToPromise<
     resolve: (returnValue: R) => void,
     reject: (error: unknown) => void
   ) => void
->(fn: T): (arg0: P0, arg1: P1) => Promise<R>;
+>(fn: T, context?: undefined | unknown): (arg0: P0, arg1: P1) => Promise<R>;
 
 export default function cognitiveServicesAsyncToPromise<
   R,
@@ -47,7 +47,7 @@ export default function cognitiveServicesAsyncToPromise<
     resolve: (returnValue: R) => void,
     reject: (error: unknown) => void
   ) => void
->(fn: T): (arg0: P0, arg1: P1, arg2: P2) => Promise<R>;
+>(fn: T, context?: undefined | unknown): (arg0: P0, arg1: P1, arg2: P2) => Promise<R>;
 
 export default function cognitiveServicesAsyncToPromise<
   R,
@@ -70,13 +70,13 @@ export default function cognitiveServicesAsyncToPromise<
     resolve: (returnValue: R) => void,
     reject: (error: unknown) => void
   ) => void
->(fn: T): (arg0: P0, arg1: P1, arg2: P2, arg3: P3) => Promise<R>;
+>(fn: T, context?: undefined | unknown): (arg0: P0, arg1: P1, arg2: P2, arg3: P3) => Promise<R>;
 
 export default function cognitiveServicesAsyncToPromise<
   R,
   T extends (...args: any[]) => void = (...args: any[]) => void
->(fn: T): (...args: Parameters<T>) => Promise<R> {
+>(fn: T, context: undefined | unknown = undefined): (...args: Parameters<T>) => Promise<R> {
   return (...args: Parameters<T>) =>
     // eslint-disable-next-line prefer-spread
-    new Promise<R>((resolve, reject) => fn.apply(undefined, [...args, resolve, reject] as unknown as Parameters<T>));
+    new Promise<R>((resolve, reject) => fn.apply(context, [...args, resolve, reject] as unknown as Parameters<T>));
 }
