+++ packages/web-speech-cognitive-services/src/SpeechServices/SpeechToText/private/FakeArray.spec.ts
